# ejlog-react-webapp/server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: EjLog WMS API
  description: |
    REST API completa per EjLog Warehouse Management System

    ## Caratteristiche
    - Gestione completa UDC (Unità Di Carico / Loading Units)
    - Gestione scomparti per ogni UDC
    - Gestione articoli e giacenze
    - Gestione liste di prelievo
    - Gestione ubicazioni magazzino
    - Autenticazione utenti

    ## Database
    - SQL Server su localhost\SQL2019
    - Database: promag

  version: 1.0.0
  contact:
    name: Ferretto Group
    email: support@ferretto.com
servers:
  - url: http://localhost:3077
    description: Development server
...
components:
  schemas:
...
    PaginationInfo:
      type: object
      properties:
        total:
          type: integer
          description: Numero totale di record
        limit:
          type: integer
          description: Numero massimo di record per pagina
        offset:
          type: integer
          description: Offset corrente
        hasMore:
          type: boolean
          description: Indica se ci sono altri record disponibili

    # Authentication Schemas
...
    Location:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [RACK, FLOOR, BUFFER, PICKING]
        aisle:
          type: string
        bay:
          type: string
        level:
          type: string
        position:
          type: string
        capacity:
          type: number
          format: float
        currentLoad:
          type: number
          format: float
        isBlocked:
          type: boolean
        warehouseId:
          type: integer
    CreateLocationRequest:
      type: object
      required:
        - code
        - type
      properties:
        code:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [RACK, FLOOR, BUFFER, PICKING]
        aisle:
          type: string
        bay:
          type: string
        level:
          type: string
        position:
          type: string
        capacity:
          type: number
        warehouseId:
          type: integer
    UpdateLocationRequest:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
        type:
          type: string
        aisle:
          type: string
        bay:
          type: string
        level:
          type: string
        position:
          type: string
        capacity:
          type: number
        isBlocked:
          type: boolean

    # Stock Schemas
...
paths:
...
  /api/locations:
    get:
      tags:
        - Locations
      summary: Lista ubicazioni
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [RACK, FLOOR, BUFFER, PICKING]
        - in: query
          name: warehouseId
          schema:
            type: integer
        - in: query
          name: isBlocked
          schema:
            type: boolean
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista ubicazioni
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Location'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'

    post:
      tags:
        - Locations
      summary: Crea nuova ubicazione
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationRequest'
      responses:
        '201':
          description: Ubicazione creata
  /api/locations/{id}:
    get:
      tags:
        - Locations
      summary: Dettaglio ubicazione
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dettaglio ubicazione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'

    put:
      tags:
        - Locations
      summary: Aggiorna ubicazione
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLocationRequest'
      responses:
        '200':
          description: Ubicazione aggiornata

    delete:
      tags:
        - Locations
      summary: Elimina ubicazione
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ubicazione eliminata

  # Stock Endpoints
...